moved indexer code into its own module, and removed some unnecessary imports
